// src/api/generated/schemas/contactRecord.ts
/**
 * Generated by orval v7.12.2 🍺
 * Do not edit manually.
 * Notifiable API
 * A simple example ASP.NET Core Web API
 * OpenAPI spec version: v1
 */
import type { ContactRecordAttributes } from './contactRecordAttributes';

export interface ContactRecord {
  id?: string;
  organizationId?: string;
  /** @nullable */
  externalId?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  androidPushToken?: string | null;
  /** @nullable */
  iosPushToken?: string | null;
  /** @nullable */
  webPushToken?: string | null;
  attributes?: ContactRecordAttributes;
}